pkg/endpoint: move endpoint map path functions to bpf.go
Signed-off by: Ian Vernon <ian@cilium.io>
